自动判分

+++ online-exam-api/src/main/java/com/exam/onlineexamapi/controller/admin/SubjectController.java
@@ -35,7 +35,7 @@ public class SubjectController {
     }
 
     @PostMapping("/delete")
-    public RestResult delete(@RequestBody Long id) {
+    public RestResult delete(@RequestBody Integer id) {
         Subject subject = subjectService.findById(id);
         return new RestResultBuilder<>().success(subjectService.delete(subject));
     }
